+++ spec/Pim/Bundle/BaseConnectorBundle/Reader/File/CsvProductReaderSpec.php
@@ -8,14 +8,14 @@ use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\CurrencyRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
-use Pim\Component\Connector\ArrayConverter\Flat\Product\Extractor\ProductAttributeFieldExtractor;
+use Pim\Component\Connector\ArrayConverter\Flat\Product\AttributeColumnInfoExtractor;
 use Prophecy\Argument;
 
 class CsvProductReaderSpec extends ObjectBehavior
 {
     function let(
         EntityManager $em,
-        ProductAttributeFieldExtractor $fieldExtractor,
+        AttributeColumnInfoExtractor $fieldExtractor,
         AttributeRepositoryInterface $attributeRepository,
         ChannelRepositoryInterface $channelRepository,
         LocaleRepositoryInterface $localeRepository,
@@ -63,10 +63,10 @@ class CsvProductReaderSpec extends ObjectBehavior
         $currencyRepository->getActivatedCurrencyCodes()->willReturn([]);
 
         $fieldInfo = ['locale_code' => null, 'scope_code' => null, 'price_currency' => null];
-        $fieldExtractor->extractAttributeFieldNameInfos('sku')->willReturn($fieldInfo);
-        $fieldExtractor->extractAttributeFieldNameInfos('name')->willReturn($fieldInfo);
-        $fieldExtractor->extractAttributeFieldNameInfos('view')->willReturn($fieldInfo);
-        $fieldExtractor->extractAttributeFieldNameInfos('manual-fr_FR')->willReturn($fieldInfo);
+        $fieldExtractor->extractColumnInfo('sku')->willReturn($fieldInfo);
+        $fieldExtractor->extractColumnInfo('name')->willReturn($fieldInfo);
+        $fieldExtractor->extractColumnInfo('view')->willReturn($fieldInfo);
+        $fieldExtractor->extractColumnInfo('manual-fr_FR')->willReturn($fieldInfo);
 
         $this->read()
             ->shouldReturn([
@@ -92,10 +92,10 @@ class CsvProductReaderSpec extends ObjectBehavior
         $localeRepository->getActivatedLocaleCodes()->willReturn(['fr']);
         $currencyRepository->getActivatedCurrencyCodes()->willReturn(['EUR']);
 
-        $fieldExtractor->extractAttributeFieldNameInfos('description-wronglocale-ecommerce')->willReturn(
+        $fieldExtractor->extractColumnInfo('description-wronglocale-ecommerce')->willReturn(
             ['locale_code' => 'wronglocale', 'scope_code' => 'ecommerce', 'price_currency' => null]
         );
-        $fieldExtractor->extractAttributeFieldNameInfos(Argument::any())->willReturn(
+        $fieldExtractor->extractColumnInfo(Argument::any())->willReturn(
             ['locale_code' => null, 'scope_code' => null, 'price_currency' => null]
         );
 
@@ -116,10 +116,10 @@ class CsvProductReaderSpec extends ObjectBehavior
         $localeRepository->getActivatedLocaleCodes()->willReturn(['fr']);
         $currencyRepository->getActivatedCurrencyCodes()->willReturn(['EUR']);
 
-        $fieldExtractor->extractAttributeFieldNameInfos('description-fr_FR-wrongscope')->willReturn(
+        $fieldExtractor->extractColumnInfo('description-fr_FR-wrongscope')->willReturn(
             ['locale_code' => 'fr', 'scope_code' => 'wrongscope', 'price_currency' => null]
         );
-        $fieldExtractor->extractAttributeFieldNameInfos(Argument::any())->willReturn(
+        $fieldExtractor->extractColumnInfo(Argument::any())->willReturn(
             ['locale_code' => null, 'scope_code' => null, 'price_currency' => null]
         );
 
@@ -140,10 +140,10 @@ class CsvProductReaderSpec extends ObjectBehavior
         $localeRepository->getActivatedLocaleCodes()->willReturn(['fr']);
         $currencyRepository->getActivatedCurrencyCodes()->willReturn(['EUR']);
 
-        $fieldExtractor->extractAttributeFieldNameInfos('price-wrongcurrency')->willReturn(
+        $fieldExtractor->extractColumnInfo('price-wrongcurrency')->willReturn(
             ['locale_code' => null, 'scope_code' => null, 'price_currency' => 'wrongcurrency']
         );
-        $fieldExtractor->extractAttributeFieldNameInfos(Argument::any())->willReturn(
+        $fieldExtractor->extractColumnInfo(Argument::any())->willReturn(
             ['locale_code' => null, 'scope_code' => null, 'price_currency' => null]
         );