+++ spec/Pim/Bundle/TransformBundle/Denormalizer/Flat/ProductValuesDenormalizerSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Bundle\TransformBundle\Denormalizer\Flat;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\TransformBundle\Denormalizer\Flat\ProductValuesDenormalizer;
-use Pim\Component\Connector\ArrayConverter\Flat\Product\Extractor\ProductAttributeFieldExtractor;
+use Pim\Component\Connector\ArrayConverter\Flat\Product\AttributeColumnInfoExtractor;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 
@@ -14,7 +14,7 @@ class ProductValuesDenormalizerSpec extends ObjectBehavior
     const FORMAT_CSV  = 'csv';
     const VALUE_CLASS = 'Pim\Bundle\CatalogBundle\Model\ProductValue';
 
-    function let(DenormalizerInterface $valueDenormalizer, ProductAttributeFieldExtractor $fieldExtractor)
+    function let(DenormalizerInterface $valueDenormalizer, AttributeColumnInfoExtractor $fieldExtractor)
     {
         $this->beConstructedWith($valueDenormalizer, $fieldExtractor, self::VALUE_CLASS);
     }
@@ -41,10 +41,10 @@ class ProductValuesDenormalizerSpec extends ObjectBehavior
 
     function it_denormalizes_variant_group_values($fieldExtractor, $valueDenormalizer, AttributeInterface $description)
     {
-        $fieldExtractor->extractAttributeFieldNameInfos('description-ecommerce-en_US')
+        $fieldExtractor->extractColumnInfo('description-ecommerce-en_US')
             ->willReturn(['attribute' => $description, 'locale_code' => 'en_US', 'scope_code' => 'ecommerce'])
             ->shouldBeCalled();
-        $fieldExtractor->extractAttributeFieldNameInfos('description-ecommerce-fr_FR')
+        $fieldExtractor->extractColumnInfo('description-ecommerce-fr_FR')
             ->willReturn(['attribute' => $description, 'locale_code' => 'fr_FR', 'scope_code' => 'ecommerce'])
             ->shouldBeCalled();